corrijo errores en el create member y sigo intentando con los tests
